backend/requirements.txt
flask
python-dotenv
PyPDF2
python-docx
openai
concurrent-log-handler